Restruct APIs in v2
