+++ pymilvus/v2/__init__.py
@@ -11,67 +11,7 @@
 # the License.
 
 # APIs that V2 provided
-from .grpc_handler import connect, insecured_connect
-
-from .types import CollectionSchema, ResourceGroupInfo
-
-from .apis import ( # alias
-    create_alias,
-    drop_alias,
-    alter_alias,
-    list_aliases,
-)
-from .apis import ( # collection
-    create_collection,
-    drop_collection,
-    has_collection,
-    describe_collection,
-    alter_collection,
-    get_collection_statistics,
-    list_collections
-)
-from .apis import ( # partition
-    create_partition,
-    drop_partition,
-    has_partition,
-    describe_partition,
-    get_partition_statistics,
-    list_partitions
-)
-from .apis import ( # index
-    drop_index,
-    has_index,
-    list_indexes,
-    describe_index,
-)
-from .apis import ( # resource group
-    create_resource_group,
-    drop_resource_group,
-    describe_resource_group,
-    list_resource_groups,
-    transfer_node,
-    transfer_replica,
-)
-from .apis import ( # others
-    get_server_version,
-)
-
-from .async_apis import ( # async APIs
-    load_collection,
-    load_partitions,
-    release_collection,
-    release_partitions,
-    flush,
-    create_index,
-    insert,
-    insert_by_rows,
-    bulk_insert,
-    upsert,
-    delete,
-    delete_by_expression,
-    search,
-    query,
-)
+from .apis import MilvusClient
 
 
 # TODO
