+++ tests/v1/test_schema.py
